Merge 115210 - Cleanup JingleInfoRequest on JingleSessionManager::Close
If NAT traversal is turned on then JingleInfoRequest adds itself as a
listener to the signal strategy.  Delete the request in
SessionManager::Close so that it removes itself as a listener and
signal strategy can be cleaned up.

BUG=108212
TEST=none


Review URL: http://codereview.chromium.org/9004024

TBR=amit@chromium.org
Review URL: http://codereview.chromium.org/9025029

git-svn-id: svn://svn.chromium.org/chrome/branches/963/src@115596 0039d316-1c4b-4281-b951-d872f2087c98